refactor(db): extract workspace and agent session metrics into db/workspace.rs
